nfo/k8s/requirements.txt
requests
django
djangorestframework
pyhelm
kubernetes
pyhelm3
six